Move "reg-alloc.c" into "jit.c"
