--- src/reg-alloc.c
@@ -1,246 +1,0 @@
-/*
- * rv32emu is freely redistributable under the MIT License. See the file
- * "LICENSE" for information on usage and redistribution of this file.
- */
-
-static int liveness[N_RV_REGS];
-
-/* The priority queue of vm registers. The one which has farthest liveness is
- * first.
- */
-static uint8_t candidate_queue[N_RV_REGS];
-
-static inline void liveness_reset()
-{
-    memset(liveness, 0xff, sizeof(liveness));
-}
-
-static inline void candidate_queue_init()
-{
-    for (int i = 0; i < N_RV_REGS; i++) {
-        candidate_queue[i] = i;
-    }
-}
-
-static int liveness_cmp(const void *l, const void *r)
-{
-    return liveness[*(uint8_t *) l] - liveness[*(uint8_t *) r];
-}
-
-/* TODO: this function could be generated by "tools/gen-jit-template.py" */
-static inline void liveness_calc(block_t *block)
-{
-    uint32_t idx;
-    rv_insn_t *ir;
-
-    /* follow the order of operator in "src/rc32_template.c" */
-    for (idx = 0, ir = block->ir_head; idx < block->n_insn;
-         idx++, ir = ir->next) {
-        switch (ir->opcode) {
-        case rv_insn_nop:
-        case rv_insn_lui:
-        case rv_insn_auipc:
-        case rv_insn_jal:
-            break;
-        case rv_insn_jalr:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_beq:
-        case rv_insn_bne:
-        case rv_insn_blt:
-        case rv_insn_bge:
-        case rv_insn_bltu:
-        case rv_insn_bgeu:
-            liveness[ir->rs1] = idx;
-            liveness[ir->rs2] = idx;
-            break;
-        case rv_insn_lb:
-        case rv_insn_lh:
-        case rv_insn_lw:
-        case rv_insn_lbu:
-        case rv_insn_lhu:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_sb:
-        case rv_insn_sh:
-        case rv_insn_sw:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_addi:
-        case rv_insn_slti:
-        case rv_insn_sltiu:
-        case rv_insn_xori:
-        case rv_insn_ori:
-        case rv_insn_andi:
-        case rv_insn_slli:
-        case rv_insn_srli:
-        case rv_insn_srai:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_add:
-        case rv_insn_sub:
-        case rv_insn_sll:
-        case rv_insn_slt:
-        case rv_insn_sltu:
-        case rv_insn_xor:
-        case rv_insn_srl:
-        case rv_insn_sra:
-        case rv_insn_or:
-        case rv_insn_and:
-            liveness[ir->rs1] = idx;
-            liveness[ir->rs2] = idx;
-            break;
-        case rv_insn_ecall:
-        case rv_insn_ebreak:
-            break;
-#if RV32_HAS(EXT_M)
-        case rv_insn_mul:
-        case rv_insn_mulh:
-        case rv_insn_mulhsu:
-        case rv_insn_mulhu:
-        case rv_insn_div:
-        case rv_insn_divu:
-        case rv_insn_rem:
-        case rv_insn_remu:
-            liveness[ir->rs1] = idx;
-            liveness[ir->rs2] = idx;
-            break;
-#endif
-#if RV32_HAS(EXT_C)
-        case rv_insn_caddi4spn:
-            liveness[rv_reg_sp] = idx;
-            break;
-        case rv_insn_clw:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_csw:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_cnop:
-            break;
-        case rv_insn_caddi:
-            liveness[ir->rd] = idx;
-            break;
-        case rv_insn_cjal:
-        case rv_insn_cli:
-        case rv_insn_clui:
-            break;
-        case rv_insn_caddi16sp:
-            liveness[ir->rd] = idx;
-            break;
-        case rv_insn_csrli:
-        case rv_insn_csrai:
-        case rv_insn_candi:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_csub:
-        case rv_insn_cxor:
-        case rv_insn_cor:
-        case rv_insn_cand:
-            liveness[ir->rs1] = idx;
-            liveness[ir->rs2] = idx;
-            break;
-        case rv_insn_cj:
-            break;
-        case rv_insn_cbeqz:
-        case rv_insn_cbnez:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_cslli:
-            liveness[ir->rd] = idx;
-            break;
-        case rv_insn_clwsp:
-            liveness[rv_reg_sp] = idx;
-            break;
-        case rv_insn_cjr:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_cmv:
-            liveness[ir->rs2] = idx;
-            break;
-        case rv_insn_cebreak:
-            break;
-        case rv_insn_cjalr:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_cadd:
-            liveness[ir->rs1] = idx;
-            liveness[ir->rs2] = idx;
-            break;
-        case rv_insn_cswsp:
-            liveness[rv_reg_sp] = idx;
-            liveness[ir->rs2] = idx;
-            break;
-#endif
-        case rv_insn_fuse1:
-            for (int i = 0; i < ir->imm2; i++) {
-                liveness[ir->fuse[i].rd] = idx;
-            }
-            break;
-        case rv_insn_fuse2:
-            liveness[ir->rs1] = idx;
-            break;
-        case rv_insn_fuse3:
-            for (int i = 0; i < ir->imm2; i++) {
-                liveness[ir->fuse[i].rs1] = idx;
-                liveness[ir->fuse[i].rs2] = idx;
-            }
-            break;
-        case rv_insn_fuse4:
-        case rv_insn_fuse7:
-            for (int i = 0; i < ir->imm2; i++) {
-                liveness[ir->fuse[i].rs1] = idx;
-            }
-            break;
-        case rv_insn_fuse5:
-        case rv_insn_fuse6:
-            break;
-        default:
-            __UNREACHABLE;
-        }
-    }
-
-    candidate_queue_init();
-    qsort(candidate_queue, N_RV_REGS, sizeof(uint8_t), liveness_cmp);
-}
-
-static inline void regs_refresh(int idx)
-{
-    for (int i = 0; i < n_host_regs; i++) {
-        if (register_map[i].vm_reg_idx == -1)
-            continue;
-        if (liveness[register_map[i].vm_reg_idx] < idx)
-            register_map[i].alive = false;
-    }
-}
-
-/* return the index in the register_map */
-static inline int reg_pick(int reserved)
-{
-    /* pick an available register */
-    for (int i = 0; i < n_host_regs; i++) {
-        if (register_map[i].reg_idx == reserved)
-            continue;
-        if (!register_map[i].alive)
-            return i;
-    }
-
-    /* If registers are exhausted, pick the one which has farthest liveness. */
-    int idx = -1;
-    for (int i = 0; i < N_RV_REGS; i++) {
-        uint8_t candidate = candidate_queue[i];
-        for (int j = 0; j < n_host_regs; j++) {
-            if (register_map[j].reg_idx == reserved)
-                continue;
-            if (register_map[j].vm_reg_idx == candidate) {
-                idx = j;
-                goto end_pick_reg;
-            }
-        }
-    }
-    __UNREACHABLE;
-
-end_pick_reg:
-    assert(idx > -1 && idx < n_host_regs);
-    return idx;
-}
